# [src] Updated cuda kernel to speedup AddDiagMatMat[kTrans, kNoTrans]