--- a/src/cudamatrix/cu-kernels.cu
+++ b/src/cudamatrix/cu-kernels.cu
@@ -6,7 +6,7 @@
 //                2013  Hainan Xu
 //                2013  Xiaohui Zhang
 //           2013-2015  Guoguo Chen
-//           2016-2017  Shiyin Kang
+//           2016-2018  Shiyin Kang
 //                2017  Hossein Hadian, Daniel Galvez
 
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -1118,7 +1118,7 @@ __global__
 static void _add_diag_mat_mat_MTN(const Real alpha, const Real* M,
                                   const int stride_M, const Real* N,
                                   const MatrixDim dim_N, const Real beta,
-                                  Real* v) {
+                                  Real* v, const int stride_v) {
   __shared__ Real ssum[CU1DBLOCK];
   const int tid = threadIdx.y * blockDim.x + threadIdx.x;
   const int j = blockIdx.x * blockDim.x + threadIdx.x;
@@ -1127,9 +1127,12 @@ static void _add_diag_mat_mat_MTN(const Real alpha, const Real* M,
     return;
 
   // Loop along the matrix column.
-  // Reduce to CU1DBLOCK / TileDim elements per column.
+  // Reduce to gridDim.y * CU1DBLOCK / TileDim elements per column.
   Real tsum = Real(0);
-  for (int i = threadIdx.y; i < dim_N.rows; i += blockDim.y) {
+
+  const int grid_stride_y = blockDim.y * gridDim.y;
+  for (int i = blockIdx.y * blockDim.y + threadIdx.y; i < dim_N.rows; i +=
+      grid_stride_y) {
     tsum += M[i * stride_M + j] * N[i * dim_N.stride + j];
   }
   ssum[tid] = tsum;
@@ -1156,7 +1159,12 @@ static void _add_diag_mat_mat_MTN(const Real alpha, const Real* M,
 
   // output TileDim sums per thread block
   if (tid < TileDim) {
-    v[j] = alpha * ssum[tid] + beta * v[j];
+    if (beta != Real(0)) {
+      v[blockIdx.y * stride_v + j] = alpha * ssum[tid]
+          + beta * v[blockIdx.y * stride_v + j];
+    } else {
+      v[blockIdx.y * stride_v + j] = alpha * ssum[tid];
+    }
   }
 }
 
@@ -4084,11 +4092,14 @@ void cudaF_add_diag_mat_mat_MNT(int Gr, int Bl, const float alpha,
 void cudaF_add_diag_mat_mat_MTN(dim3 Gr, dim3 Bl, const float alpha,
                                 const float* M, const int stride_M,
                                 const float* N, const MatrixDim dim_N,
-                                const float beta, float* v) {
+                                const float beta, float* v,
+                                const int stride_v) {
   if (Bl.x == 16) {
-    _add_diag_mat_mat_MTN<16> <<<Gr,Bl>>>(alpha,M,stride_M,N,dim_N,beta,v);
-  } else if (Bl.x==32) {
-    _add_diag_mat_mat_MTN<32><<<Gr,Bl>>>(alpha,M,stride_M,N,dim_N,beta,v);
+    _add_diag_mat_mat_MTN<16> <<<Gr, Bl>>>(alpha, M, stride_M, N, dim_N, beta,
+                                           v, stride_v);
+  } else if (Bl.x == 32) {
+    _add_diag_mat_mat_MTN<32> <<<Gr, Bl>>>(alpha, M, stride_M, N, dim_N, beta,
+                                           v, stride_v);
   }
 }
 
@@ -4781,11 +4792,14 @@ void cudaD_add_diag_mat_mat_MNT(int Gr, int Bl, const double alpha,
 void cudaD_add_diag_mat_mat_MTN(dim3 Gr, dim3 Bl, const double alpha,
                                 const double* M, const int stride_M,
                                 const double* N, const MatrixDim dim_N,
-                                const double beta, double* v) {
+                                const double beta, double* v,
+                                const int stride_v) {
   if (Bl.x == 16) {
-    _add_diag_mat_mat_MTN<16> <<<Gr,Bl>>>(alpha,M,stride_M,N,dim_N,beta,v);
-  } else if (Bl.x==32) {
-    _add_diag_mat_mat_MTN<32><<<Gr,Bl>>>(alpha,M,stride_M,N,dim_N,beta,v);
+    _add_diag_mat_mat_MTN<16> <<<Gr, Bl>>>(alpha, M, stride_M, N, dim_N, beta,
+                                           v, stride_v);
+  } else if (Bl.x == 32) {
+    _add_diag_mat_mat_MTN<32> <<<Gr, Bl>>>(alpha, M, stride_M, N, dim_N, beta,
+                                           v, stride_v);
   }
 }