--- a/src/cudamatrix/cu-vector-speed-test.cc
+++ b/src/cudamatrix/cu-vector-speed-test.cc
@@ -1,7 +1,9 @@
 // cudamatrix/cu-vector-speed-test.cc
 
-// Copyright 2013  Johns Hopkins University (author: Daniel Povey)
-// Copyright 2017  Daniel Galvez
+// Copyright      2013  Johns Hopkins University (author: Daniel Povey)
+//                2017  Daniel Galvez
+//           2016-2018  Shiyin Kang
+
 
 // See ../../COPYING for clarification regarding multiple authors
 //
@@ -211,7 +213,8 @@ template<typename Real> void TestCuVectorAddDiagMat2OnVariousShapes(
   BaseFloat gflops = (fdim * iter) / (tim.Elapsed() * 1.0e+09);
   KALDI_LOG << "For CuVector::AddDiagMat2Shapes" << NameOf<Real>()
             << (trans == kTrans ? "[trans]" : "[no-trans]") << ", for dim = ("
-            << size / dim << ", " << dim  << "), speed was " << gflops << " gigaflops.";
+            << size / dim << ", " << dim  << "), speed was " << gflops
+            << " gigaflops.";
 }
 
 
@@ -366,13 +369,54 @@ template<typename Real> void TestCuVectorApplyCeilingNoCount(int32 dim) {
 }
 
 
+template<typename Real> void TestCuVectorAddDiagMatMatShape(
+    int32 num_rows, int32 num_cols, MatrixTransposeType transM,
+    MatrixTransposeType transN) {
+  BaseFloat time_in_secs = 0.02;
+  CuVector<Real> v(transM == kTrans ? num_cols : num_rows);
+  v.SetRandn();
+  CuMatrix<Real> M(num_rows, num_cols);
+  CuMatrix<Real> N(transM != transN ? num_rows : num_cols,
+                   transM != transN ? num_cols : num_rows);
+  M.SetRandn();
+  N.SetRandn();
+
+  Timer tim;
+  int32 iter = 0;
+
+  for (;tim.Elapsed() < time_in_secs; iter++) {
+    v.AddDiagMatMat(1.0, M, transM, N, transN, 1.0);
+  }
+
+  BaseFloat fnr = num_rows;
+  BaseFloat fnc = num_cols;
+  BaseFloat gflops = (fnr * fnc * iter) / (tim.Elapsed() * 1.0e+09);
+  KALDI_LOG << "For CuVector::AddDiagMatMat" << NameOf<Real>()
+            << (transM == kNoTrans ? "[no-trans],":"[trans],")
+            << (transN == kNoTrans ? "[no-trans],":"[trans],")
+            << " for dim = "<< num_rows << ", " << num_cols
+            << ", speed was " << gflops << " gigaflops.";
+}
+
+
 template<typename Real> void CudaVectorSpeedTest() {
+  const size_t a = 1 << 5;
+  const size_t b = 1 << 8;
+  for (size_t i = a; i <= b; i *= 2) {
+    for (size_t j = a; j <= b; j *= 2) {
+      if (i * j <= a * b) {
+        TestCuVectorAddDiagMatMatShape<Real>(i, j, kNoTrans, kNoTrans);
+        TestCuVectorAddDiagMatMatShape<Real>(i, j, kNoTrans, kTrans);
+        TestCuVectorAddDiagMatMatShape<Real>(i, j, kTrans, kNoTrans);
+        TestCuVectorAddDiagMatMatShape<Real>(i, j, kTrans, kTrans);
+      }
+    }
+  }
+
   std::vector<int32> sizes;
-  sizes.push_back(32);
-  sizes.push_back(64);
-  sizes.push_back(128);
-  sizes.push_back(256);
-  sizes.push_back(1024);
+  for (int i = 32; i <= 1024; i *= 2) {
+    sizes.push_back(i);
+  }
   int32 ns = sizes.size();
   for (int32 s = 0; s < ns; s++)
     TestCuVectorSoftmax<Real>(sizes[s]);
@@ -443,3 +487,4 @@ int main() {
 #endif
   KALDI_LOG << "Tests succeeded.";
 }
+