This is a further improvement for #2552 .

It uses 2D grid on large but thin matrix to fully utilize the GPU, with the cost of adding an additional kernel for a second stage reduction. The overall speed on matrix of the size (1048576, 32) is ~52gflops, which is close to the possible peak of ~56 gflops.

All GPU tests in `cudamatrix` and `nnet*` have passed.

```
Speedup(ratio) <float> (num_rows, num_cols)
            32     64    128    256    512   1024   2048   4096   8192  16384  32768  65536 131072 262144 524288 1048576
     32   1.07   1.08   1.09   1.09   1.09   1.07   1.09   0.98   1.01   1.06   1.07   1.10   1.06   1.07   1.05   1.06
     64   1.03   1.03   1.04   1.03   1.01   1.02   1.05   0.95   1.02   1.02   1.03   1.05   1.05   1.02   1.05
    128   1.01   1.01   1.02   1.03   1.01   1.01   1.11   0.93   1.05   1.01   1.03   1.01   1.01   1.03
    256   1.01   1.01   1.01   1.01   1.03   1.01   1.19   0.97   1.09   1.01   1.02   1.01   1.03
    512   1.02   1.02   1.02   1.02   1.02   1.02   1.30   0.99   1.09   1.01   1.01   1.00
   1024   1.02   1.00   1.02   1.03   1.02   1.02   1.35   1.00   1.11   1.01   1.01
   2048   1.02   1.02   1.02   1.02   1.03   1.02   1.39   0.99   1.09   1.01
   4096   1.07   1.08   1.11   1.30   1.16   0.95   1.15   0.93   1.09
   8192   1.98   2.01   2.41   2.07   1.70   1.22   1.35   1.01
  16384   0.63   0.97   1.26   1.66   2.20   1.47   1.51
  32768   0.91   1.22   1.68   2.06   2.65   1.66
  65536   1.28   1.67   2.06   2.59   2.95
 131072   1.90   2.28   2.62   3.38
 262144   2.37   2.62   3.34
 524288   2.80   3.37
1048576   3.17

Speedup(ratio) <double> (num_rows, num_cols)
            32     64    128    256    512   1024   2048   4096   8192  16384  32768  65536 131072 262144 524288 1048576
     32   0.99   0.99   1.00   1.01   1.00   0.99   1.01   1.02   1.01   1.16   1.16   1.20   1.24   1.25   1.18   1.25
     64   0.99   0.99   0.99   1.01   1.00   0.99   1.00   1.02   1.01   1.09   1.08   1.10   1.13   1.10   1.11
    128   0.99   0.96   0.99   1.00   1.01   1.00   1.00   1.01   1.02   1.06   1.05   1.06   1.05   1.06
    256   1.00   1.00   1.01   1.00   1.00   0.99   1.01   1.00   1.01   1.03   1.01   1.04   1.04
    512   1.00   1.00   1.00   1.00   1.00   1.01   1.00   1.00   1.01   1.02   1.00   1.02
   1024   1.00   1.00   1.00   1.00   1.01   1.01   1.00   1.00   1.00   1.02   1.01
   2048   1.00   1.00   1.00   1.00   1.01   1.00   1.00   1.00   1.00   1.02
   4096   1.26   1.29   1.55   1.31   1.08   0.87   0.82   0.97   0.96
   8192   2.21   2.81   2.41   1.87   1.38   1.01   0.91   1.03
  16384   0.87   1.24   1.70   2.18   1.65   1.11   0.95
  32768   1.25   1.68   2.18   2.56   1.83   1.16
  65536   1.70   2.14   2.56   3.04   1.92
 131072   2.36   2.63   3.05   4.59
 262144   2.96   3.24   4.12
 524288   3.23   4.11
1048576   3.46

new(gflops) <float> (num_rows, num_cols)
            32     64    128    256    512   1024   2048   4096   8192  16384  32768  65536 131072 262144 524288 1048576
     32   0.09   0.17   0.35   0.70   1.38   2.69   5.22   9.02  15.13  25.55  30.48  38.59  43.75  46.10  49.84  50.88
     64   0.17   0.34   0.67   1.34   2.61   5.11   9.76  16.02  25.43  32.60  39.81  46.53  50.99  50.94  55.38
    128   0.31   0.61   1.23   2.45   4.77   9.34  17.38  25.90  31.78  41.79  47.61  52.26  53.54  55.71
    256   0.53   1.05   2.11   4.21   8.31  15.88  28.72  30.89  41.12  49.26  52.71  54.86  57.29
    512   0.83   1.66   3.32   6.64  12.97  24.89  31.67  38.25  47.39  53.83  55.37  56.55
   1024   1.16   2.28   4.63   9.30  18.06  24.38  38.76  43.81  51.76  56.44  56.73
   2048   1.44   2.89   5.78  11.60  15.37  28.48  44.00  47.26  53.15  57.90
   4096   1.73   3.49   6.91  11.26  19.00  28.91  38.11  46.70  52.98
   8192   3.41   6.91  11.20  18.90  29.08  38.22  45.93  52.17
  16384   6.79  12.04  19.05  28.99  38.58  46.70  51.86
  32768  11.21  19.03  29.08  38.43  46.96  52.38
  65536  18.85  29.21  38.51  47.06  52.73
 131072  29.00  39.17  47.02  52.99
 262144  39.53  47.53  52.75
 524288  48.01  52.87
1048576  52.85

new(gflops) <double> (num_rows, num_cols)
            32     64    128    256    512   1024   2048   4096   8192  16384  32768  65536 131072 262144 524288 1048576
     32   0.08   0.17   0.33   0.67   1.29   2.51   4.70   7.72  12.15  16.30  21.24  25.36  28.07  29.38  28.91  31.23
     64   0.16   0.31   0.63   1.26   2.45   4.72   8.40  12.69  16.18  21.78  25.58  28.73  30.38  31.32  31.89
    128   0.29   0.55   1.15   2.28   4.44   8.42  13.71  16.13  21.65  26.10  28.62  30.75  31.47  32.07
    256   0.49   0.97   1.96   3.85   7.46  13.78  17.95  20.62  25.89  28.91  30.20  31.83  32.17
    512   0.74   1.49   3.00   5.87  11.26  15.87  23.01  24.04  28.83  30.60  31.11  32.31
   1024   1.01   2.03   4.09   7.99  11.51  19.55  26.74  26.22  30.52  31.46  31.49
   2048   1.24   2.47   5.00   7.22  13.42  22.23  29.05  27.60  31.33  31.89
   4096   1.75   3.55   6.18  10.33  15.63  20.83  25.05  27.14  30.17
   8192   3.26   5.85  10.06  15.39  20.97  25.40  28.31  29.27
  16384   5.61   9.93  15.30  21.04  25.51  28.41  30.01
  32768   9.70  15.34  21.01  25.47  28.48  30.23
  65536  15.03  20.78  25.47  28.40  30.23
 131072  20.41  24.10  28.56  30.30
 262144  24.82  28.20  30.37
 524288  27.79  30.10
1048576  29.76

old(gflops) <float> (num_rows, num_cols)
            32     64    128    256    512   1024   2048   4096   8192  16384  32768  65536 131072 262144 524288 1048576
     32   0.08   0.16   0.32   0.64   1.27   2.52   4.79   9.19  14.96  24.12  28.57  35.03  41.17  43.11  47.37  48.12
     64   0.16   0.33   0.65   1.30   2.58   5.03   9.27  16.95  25.05  31.83  38.64  44.26  48.41  49.84  52.92
    128   0.30   0.61   1.20   2.38   4.73   9.24  15.63  27.82  30.31  41.21  46.43  51.55  52.91  53.84
    256   0.52   1.04   2.08   4.15   8.09  15.69  24.11  31.76  37.79  48.63  51.85  54.15  55.79
    512   0.81   1.63   3.26   6.51  12.69  24.29  24.41  38.72  43.39  53.39  55.01  56.44
   1024   1.14   2.27   4.54   9.06  17.67  23.96  28.74  43.91  46.77  56.14  56.35
   2048   1.42   2.83   5.67  11.34  14.94  27.97  31.56  47.59  48.55  57.48
   4096   1.62   3.22   6.21   8.67  16.42  30.48  33.10  50.01  48.64
   8192   1.72   3.44   4.64   9.12  17.12  31.45  34.01  51.60
  16384  10.76  12.40  15.17  17.51  17.52  31.67  34.43
  32768  12.34  15.59  17.32  18.65  17.74  31.52
  65536  14.76  17.46  18.69  18.17  17.86
 131072  15.28  17.21  17.94  15.68
 262144  16.70  18.12  15.79
 524288  17.17  15.70
1048576  16.69

old(gflops) <double> (num_rows, num_cols)
            32     64    128    256    512   1024   2048   4096   8192  16384  32768  65536 131072 262144 524288 1048576
     32   0.08   0.17   0.33   0.66   1.30   2.54   4.67   7.57  12.01  14.08  18.24  21.19  22.56  23.54  24.52  24.92
     64   0.16   0.32   0.63   1.25   2.46   4.76   8.40  12.40  15.95  20.04  23.59  26.23  26.99  28.36  28.64
    128   0.29   0.58   1.16   2.27   4.41   8.44  13.77  15.94  21.18  24.60  27.22  28.97  29.95  30.25
    256   0.49   0.97   1.95   3.83   7.45  13.89  17.75  20.54  25.72  27.94  29.78  30.62  30.89
    512   0.75   1.49   2.99   5.88  11.24  15.73  22.92  23.94  28.68  29.89  30.95  31.53
   1024   1.01   2.03   4.09   8.01  11.36  19.44  26.72  26.24  30.41  30.79  31.31
   2048   1.24   2.48   5.00   7.19  13.24  22.21  29.06  27.48  31.26  31.13
   4096   1.39   2.76   3.98   7.86  14.46  24.05  30.40  28.04  31.46
   8192   1.47   2.08   4.18   8.25  15.14  25.13  31.06  28.43
  16384   6.48   8.01   9.01   9.63  15.45  25.71  31.53
  32768   7.77   9.10   9.64   9.95  15.59  26.14
  65536   8.82   9.72   9.95   9.35  15.72
 131072   8.66   9.15   9.36   6.60
 262144   8.39   8.70   7.38
 524288   8.60   7.32
1048576   8.60
```