@danpovey 2018-07-20
Fantastic!  Thanks!
I'll merge when the checks complete.